@@ -4,7 +4,7 @@ build-backend = "hatchling.build"
 
 [project]
 name = "camel-ai"
-version = "0.2.31"
+version = "0.2.36"
 description = "Communicative Agents for AI Society Study"
 authors = [{ name = "CAMEL-AI.org" }]
 requires-python = ">=3.10,<3.13"
@@ -26,11 +26,8 @@ dependencies = [
     "tiktoken>=0.7.0,<0.8",
     "colorama>=0.4.6,<0.5",
     "jsonschema>=4,<5",
-    "protobuf>=5,<6",
     "docstring-parser>=0.15,<0.16",
     "pydantic>=1.9,<2.10",
-    "eval-type-backport==0.2.0",
-    "curl_cffi==0.6.2",
     "httpx>=0.28.0,<1.0.0dev",
     "psutil>=5.9.8,<6",
     "pyyaml>=6.0.2,<7",
@@ -182,6 +179,7 @@ storage = [
     "azure-storage-blob>=12.21.0,<13",
     "google-cloud-storage>=2.18.0,<3",
     "botocore>=1.35.3,<2",
+    "mem0ai>=0.1.73"
 ]
 all = [
     "transformers>=4,<5",
@@ -283,6 +281,7 @@ all = [
     "fpdf>=1.7.2",
     "mcp>=1.3.0",
     "typer>=0.15.2",
+    "mem0ai>=0.1.67"
 ]
 
 [project.urls]
@@ -305,7 +304,7 @@ include = ["camel"]
 [tool.ruff]
 line-length = 79
 fix = true
-target-version = "py39"
+target-version = "py310"
 
 [tool.ruff.format]
 quote-style = "preserve"
@@ -459,5 +458,6 @@ module = [
     "fpdf",
     "mcp",
     "typer",
+    "mem0",
 ]
 ignore_missing_imports = true
