@@ -29,19 +29,19 @@ class DeepSeekConfig(BaseConfig):
         temperature (float, optional): Sampling temperature to use, between
             :obj:`0` and :obj:`2`. Higher values make the output more random,
             while lower values make it more focused and deterministic.
-            (default: :obj:`1.0`)
+            (default: :obj:`None`)
         top_p (float, optional): Controls the diversity and focus of the
             generated results. Higher values make the output more diverse,
-            while lower values make it more focused. (default: :obj:`1.0`)
+            while lower values make it more focused. (default: :obj:`None`)
         response_format (object, optional): Specifies the format of the
             returned content. The available values are `{"type": "text"}` or
             `{"type": "json_object"}`. Setting it to `{"type": "json_object"}`
             will output a standard JSON string.
-            (default: :obj:`{"type": "text"}`)
+            (default: :obj:`None`)
         stream (bool, optional): If set, partial message deltas will be sent.
             Tokens will be sent as data-only server-sent events (SSE) as
             they become available, with the stream terminated by a
-            data: [DONE] message. (default: :obj:`False`)
+            data: [DONE] message. (default: :obj:`None`)
         stop (Union[str, list[str]], optional): Up to 16 sequences where
             the API will stop generating further tokens. (default: :obj:`None`)
         max_tokens (int, optional): The maximum number of tokens that can
@@ -51,11 +51,11 @@ class DeepSeekConfig(BaseConfig):
         presence_penalty (float, optional): Number between -2.0 and 2.0.
             Positive values penalize new tokens based on whether they
             appear in the text so far, increasing the model's likelihood
-            to talk about new topics. (default: :obj:`0.0`)
+            to talk about new topics. (default: :obj:`None`)
         frequency_penalty (float, optional): Number between -2.0 and 2.0.
             Positive values penalize new tokens based on their existing
             frequency in the text so far, decreasing the model's likelihood
-            to repeat the same line verbatim. (default: :obj:`0`)
+            to repeat the same line verbatim. (default: :obj:`None`)
         tools (list[FunctionTool], optional): A list of tools the model may
             call. Currently, only functions are supported as a tool. Use
             this to provide a list of functions the model may generate JSON
@@ -70,31 +70,31 @@ class DeepSeekConfig(BaseConfig):
             {"type": "function", "function": {"name": "my_function"}} forces
             the model to call that tool. "none" is the default when no tools
             are present. "auto" is the default if tools are present.
-            (default: :obj:`"auto"`)
+            (default: :obj:`None`)
         logprobs (bool, optional): Whether to return log probabilities of
             the output tokens or not. If true, returns the log probabilities
             of each output token returned in the content of message.
-            (default: :obj:`False`)
+            (default: :obj:`None`)
         top_logprobs (int, optional): An integer between 0 and 20 specifying
             the number of most likely tokens to return at each token
             position, each with an associated log probability. logprobs
             must be set to true if this parameter is used.
             (default: :obj:`None`)
         include_usage (bool, optional): When streaming, specifies whether to
-            include usage information in `stream_options`. (default:
-            :obj:`True`)
+            include usage information in `stream_options`.
+            (default: :obj:`None`)
     """
 
-    temperature: float = 1.0  # deepseek default: 1.0
-    top_p: float = 1.0
-    stream: bool = False
+    temperature: Optional[float] = None  # deepseek default: 1.0
+    top_p: Optional[float] = None
+    stream: Optional[bool] = None
     stop: Optional[Union[str, Sequence[str]]] = None
     max_tokens: Optional[int] = None
-    presence_penalty: float = 0.0
+    presence_penalty: Optional[float] = None
     response_format: Optional[Union[Type[BaseModel], dict]] = None
-    frequency_penalty: float = 0.0
+    frequency_penalty: Optional[float] = None
     tool_choice: Optional[Union[dict[str, str], str]] = None
-    logprobs: bool = False
+    logprobs: Optional[bool] = None
     top_logprobs: Optional[int] = None
 
     def __init__(self, include_usage: bool = True, **kwargs):
