@@ -16,7 +16,6 @@ from __future__ import annotations
 from typing import Any, Dict, List, Optional, Sequence, Union
 
 from camel.configs.base_config import BaseConfig
-from camel.types import NOT_GIVEN, NotGiven
 
 
 class SGLangConfig(BaseConfig):
@@ -31,27 +30,27 @@ class SGLangConfig(BaseConfig):
         temperature (float, optional): Sampling temperature to use, between
             :obj:`0` and :obj:`2`. Higher values make the output more random,
             while lower values make it more focused and deterministic.
-            (default: :obj:`1.0`)
+            (default: :obj:`None`)
         top_p (float, optional): An alternative to sampling with temperature,
             called nucleus sampling, where the model considers the results of
             the tokens with top_p probability mass. So :obj:`0.1` means only
             the tokens comprising the top 10% probability mass are considered.
-            (default: :obj:`1.0`)
+            (default: :obj:`None`)
         n (int, optional): How many chat completion choices to generate for
-            each input message. (default: :obj:`1`)
+            each input message. (default: :obj:`None`)
         frequency_penalty (float, optional): Number between :obj:`-2.0` and
             :obj:`2.0`. Positive values penalize new tokens based on their
             existing frequency in the text so far, decreasing the model's
             likelihood to repeat the same line verbatim. See more information
-            about frequency and presence penalties. (default: :obj:`0.0`)
+            about frequency and presence penalties. (default: :obj:`None`)
         presence_penalty (float, optional): Number between :obj:`-2.0` and
             :obj:`2.0`. Positive values penalize new tokens based on whether
             they appear in the text so far, increasing the model's likelihood
             to talk about new topics. See more information about frequency and
-            presence penalties. (default: :obj:`0.0`)
+            presence penalties. (default: :obj:`None`)
         stream (bool, optional): Whether to stream the generated output in
             chunks. If set to `True`, the response will be streamed as it is
-            generated. (default: :obj:`False`)
+            generated. (default: :obj:`None`)
         max_tokens (int, optional): The maximum number of tokens to generate
             in the chat completion. The total length of input tokens and
             generated tokens is limited by the model's context length.
@@ -60,17 +59,17 @@ class SGLangConfig(BaseConfig):
             that the model can dynamically invoke. Each tool should be
             defined as a dictionary following OpenAI's function calling
             specification format. For more details, refer to the OpenAI
-            documentation.
+            documentation. (default: :obj:`None`)
     """
 
-    stop: Union[str, Sequence[str], NotGiven] = NOT_GIVEN
-    temperature: float = 1.0
-    top_p: float = 1.0
-    n: int = 1
-    frequency_penalty: float = 0.0
-    presence_penalty: float = 0.0
-    stream: bool = False
-    max_tokens: Union[int, NotGiven] = NOT_GIVEN
+    stop: Optional[Union[str, Sequence[str]]] = None
+    temperature: Optional[float] = None
+    top_p: Optional[float] = None
+    n: Optional[int] = None
+    frequency_penalty: Optional[float] = None
+    presence_penalty: Optional[float] = None
+    stream: Optional[bool] = None
+    max_tokens: Optional[int] = None
     tools: Optional[Union[List[Dict[str, Any]]]] = None
 
 
