@@ -110,9 +110,21 @@ class GithubToolkit(BaseToolkit):
                 successfully or not.
         """
         sb = self.repo.get_branch(self.repo.default_branch)
-        self.repo.create_git_ref(
-            ref=f"refs/heads/{branch_name}", sha=sb.commit.sha
-        )
+        from github import GithubException
+
+        try:
+            self.repo.create_git_ref(
+                ref=f"refs/heads/{branch_name}", sha=sb.commit.sha
+            )
+        except GithubException as e:
+            if e.message == "Reference already exists":
+                # agent might have pushed the branch separately.
+                logger.warning(
+                    f"Branch {branch_name} already exists. "
+                    "Continuing with the existing branch."
+                )
+            else:
+                raise
 
         file = self.repo.get_contents(file_path)
 
