@@ -12,12 +12,11 @@
 # limitations under the License.
 # ========= Copyright 2023-2024 @ CAMEL-AI.org. All Rights Reserved. =========
 from .base import BaseVerifier
-from .models import VerificationOutcome, VerifierInput
+from .models import VerificationOutcome
 from .python_verifier import PythonVerifier
 
 __all__ = [
     "BaseVerifier",
     "VerificationOutcome",
-    "VerifierInput",
     "PythonVerifier",
 ]
