@@ -13,10 +13,10 @@
 # ========= Copyright 2023-2024 @ CAMEL-AI.org. All Rights Reserved. =========
 import asyncio
 
-from camel.verifiers import PythonVerifier, VerifierInput
+from camel.verifiers import PythonVerifier
 
 verifier = PythonVerifier(required_packages=["numpy"])
-asyncio.run(verifier.setup())
+asyncio.run(verifier.setup(uv=True))
 
 numpy_test_code = """
 import numpy as np
@@ -29,15 +29,17 @@ print(result)
 
 # Since the output of the above numpy code evaluates to 32,
 # we expect the verification outcome to be a success.
-response = VerifierInput(llm_response=numpy_test_code, ground_truth="32")
-result = asyncio.run(verifier.verify(response))
-print(f"Result: {result.status}")
+result = asyncio.run(
+    verifier.verify(solution=numpy_test_code, ground_truth="32")
+)
+print(f"Result: {result}")
 
-response = VerifierInput(llm_response=numpy_test_code, ground_truth="40")
-result = asyncio.run(verifier.verify(response))
+result = asyncio.run(
+    verifier.verify(solution=numpy_test_code, ground_truth="40")
+)
 
 # Now we expect the VerificationOutcome to be a failure,
 # because the answer is wrong.
-print(f"Result: {result.status}")
+print(f"Result: {result}")
 
 asyncio.run(verifier.cleanup())
