@@ -27,6 +27,9 @@ from camel.datasets import (
     FewShotGenerator,
     StaticDataset,
 )
+from camel.models.base_model import BaseModelBackend
+from camel.models.stub_model import StubModel
+from camel.verifiers import BaseVerifier
 
 
 # ruff: noqa: RUF001
@@ -165,13 +168,11 @@ def test_static_dataset_init_from_hf_dataset():
 
     # **Test 2: Initialization with invalid data**
     # Sub-test 2a: Missing required field with strict=True
-    invalid_data_missing = [
-        {"question": "What is 3 + 3?", "final_answer": "6"}
-    ]
+    invalid_data_missing = [{"question": "What is 3 + 3?"}]
     hf_invalid_missing = HFDataset.from_list(invalid_data_missing)
     with pytest.raises(
         ValueError,
-        match="Sample at index 0 has invalid 'rationale': "
+        match="Sample at index 0 has invalid 'final_answer': "
         "expected str, got <class 'NoneType'>",
     ):
         StaticDataset(data=hf_invalid_missing, min_samples=1, strict=True)
@@ -297,13 +298,12 @@ def test_static_dataset_init_from_pytorch_dataset():
     invalid_data_missing = [
         {
             "question": "What is 3 + 3?",
-            "final_answer": "6",
-        }  # Missing rationale
+        }  # Missing final_answer
     ]
     pytorch_invalid_missing = MockPyTorchDataset(invalid_data_missing)
     with pytest.raises(
         ValueError,
-        match="Sample at index 0 has invalid 'rationale': "
+        match="Sample at index 0 has invalid 'final_answer': "
         "expected str, got <class 'NoneType'>",
     ):
         StaticDataset(data=pytorch_invalid_missing, min_samples=1, strict=True)
@@ -564,15 +564,14 @@ def test_static_dataset_init_from_json_file():
         invalid_data = [
             {
                 "question": "What is 3 + 3?",
-                "final_answer": "6",
-            }  # Missing rationale
+            }  # Missing final answer
         ]
         json.dump(invalid_data, tmp_file)
         tmp_file_path = Path(tmp_file.name)
 
     with pytest.raises(
         ValueError,
-        match="Sample at index 0 has invalid 'rationale': "
+        match="Sample at index 0 has invalid 'final_answer': "
         "expected str, got <class 'NoneType'>",
     ):
         StaticDataset(data=tmp_file_path, min_samples=1, strict=True)
@@ -584,10 +583,211 @@ def test_static_dataset_init_from_json_file():
         invalid_data = [
             {
                 "question": "What is 3 + 3?",
-                "final_answer": "6",
-            }  # Missing rationale
+            }  # Missing final_answer
         ]
         json.dump(invalid_data, tmp_file)
+        tmp_file_path = Path(tmp_file.name)
+
+    dataset_non_strict = StaticDataset(
+        data=tmp_file_path, min_samples=0, strict=False
+    )
+    assert (
+        len(dataset_non_strict) == 0
+    ), "Invalid items should be filtered out in non-strict mode."
+
+
+def test_static_dataset_init_from_jsonl_file():
+    r"""
+    Test StaticDataset initialization from a JSONL file with various scenarios.
+    """
+    import json
+    import tempfile
+    from pathlib import Path
+
+    import pytest
+
+    # **Test 1: Initialization with a valid JSONL file (only required fields)**
+    with tempfile.NamedTemporaryFile(
+        mode='w', suffix='.jsonl', delete=False
+    ) as tmp_file:
+        valid_data = [
+            {
+                "question": "What is 2 + 2?",
+                "rationale": "Addition of 2 and 2.",
+                "final_answer": "4",
+            },
+            {
+                "question": "Is the sky blue?",
+                "rationale": "Observation of clear weather.",
+                "final_answer": "yes",
+            },
+        ]
+        # Write each dictionary as a separate JSON line.
+        for item in valid_data:
+            tmp_file.write(json.dumps(item) + "\n")
+        tmp_file_path = Path(tmp_file.name)
+
+    dataset = StaticDataset(data=tmp_file_path, min_samples=1, strict=True)
+    assert len(dataset) == 2, "Dataset should contain 2 items."
+    assert isinstance(
+        dataset[0], DataPoint
+    ), "Items should be DataPoint instances."
+    assert (
+        dataset[0].question == "What is 2 + 2?"
+    ), "Question should match input."
+    assert (
+        dataset[0].rationale == "Addition of 2 and 2."
+    ), "Rationale should match input."
+    assert dataset[0].final_answer == "4", "Final answer should match input."
+    assert (
+        dataset[1].question == "Is the sky blue?"
+    ), "Second question should match input."
+    assert (
+        dataset[1].rationale == "Observation of clear weather."
+    ), "Second rationale should match input."
+    assert (
+        dataset[1].final_answer == "yes"
+    ), "Second final answer should match input."
+    assert (
+        dataset[0].metadata is None
+    ), "Metadata should be None when not provided."
+
+    # **Test 2: Initialization with invalid JSONL file (malformed JSON)**
+    with tempfile.NamedTemporaryFile(
+        mode='w', suffix='.jsonl', delete=False
+    ) as tmp_file:
+        # Write a malformed JSON line (missing closing brace).
+        tmp_file.write(
+            '{"question": "Test", "rationale": "Test",'
+            '"final_answer": "Test"\n'
+        )
+        tmp_file_path = Path(tmp_file.name)
+
+    with pytest.raises(ValueError, match="Invalid JSON on line"):
+        StaticDataset(data=tmp_file_path, min_samples=1, strict=True)
+
+    # **Test 3: Initialization with JSONL file containing non-dict items**
+    with tempfile.NamedTemporaryFile(
+        mode='w', suffix='.jsonl', delete=False
+    ) as tmp_file:
+        tmp_file.write(
+            json.dumps(
+                {
+                    "question": "Test",
+                    "rationale": "Test",
+                    "final_answer": "Test",
+                }
+            )
+            + "\n"
+        )
+        # Write a non-dict item on the second line.
+        tmp_file.write(json.dumps("not a dict") + "\n")
+        tmp_file_path = Path(tmp_file.name)
+
+    with pytest.raises(ValueError, match="Expected a dictionary at record 2"):
+        StaticDataset(data=tmp_file_path, min_samples=1, strict=True)
+
+    # **Test 4: Initialization with a missing JSONL file**
+    missing_file_path = Path("non_existent_file.jsonl")
+    with pytest.raises(FileNotFoundError, match="JSONL file not found:"):
+        StaticDataset(data=missing_file_path, min_samples=1, strict=True)
+
+    # **Test 5: Initialization with an empty JSONL file**
+    with tempfile.NamedTemporaryFile(
+        mode='w', suffix='.jsonl', delete=False
+    ) as tmp_file:
+        tmp_file_path = Path(tmp_file.name)
+
+    # Sub-test 5a: Empty dataset with min_samples=1
+    with pytest.raises(
+        ValueError, match="The dataset does not contain enough samples"
+    ):
+        StaticDataset(data=tmp_file_path, min_samples=1, strict=True)
+
+    # Sub-test 5b: Empty dataset with min_samples=0
+    dataset_empty = StaticDataset(
+        data=tmp_file_path, min_samples=0, strict=True
+    )
+    assert (
+        len(dataset_empty) == 0
+    ), "Empty dataset should have zero length when min_samples=0."
+
+    # **Test 6: Initialization with optional fields and additional fields**
+    with tempfile.NamedTemporaryFile(
+        mode='w', suffix='.jsonl', delete=False
+    ) as tmp_file:
+        data_with_optional = [
+            {
+                "question": "What is the capital of France?",
+                "rationale": "France is a country in Europe.",
+                "final_answer": "Paris",
+                "metadata": {"source": "geography", "id": 123},
+                "extra_field": "this should be ignored",
+            }
+        ]
+        for item in data_with_optional:
+            tmp_file.write(json.dumps(item) + "\n")
+        tmp_file_path = Path(tmp_file.name)
+
+    dataset_optional = StaticDataset(
+        data=tmp_file_path, min_samples=1, strict=True
+    )
+    assert len(dataset_optional) == 1, "Dataset should contain 1 item."
+    assert isinstance(
+        dataset_optional[0], DataPoint
+    ), "Item should be a DataPoint instance."
+    assert (
+        dataset_optional[0].question == "What is the capital of France?"
+    ), "Question should match."
+    assert (
+        dataset_optional[0].rationale == "France is a country in Europe."
+    ), "Rationale should match."
+    assert (
+        dataset_optional[0].final_answer == "Paris"
+    ), "Final answer should match."
+    assert dataset_optional[0].metadata == {
+        "source": "geography",
+        "id": 123,
+    }, "Metadata should match."
+    assert (
+        "extra_field" not in dataset_optional[0].__dict__
+    ), "Extra field should not be present in DataPoint."
+
+    # **Test 7: JSONL-specific edge cases:
+    # Missing required field in strict mode**
+    with tempfile.NamedTemporaryFile(
+        mode='w', suffix='.jsonl', delete=False
+    ) as tmp_file:
+        # Missing 'final_answer' field.
+        tmp_file.write(
+            json.dumps(
+                {
+                    "question": "What is 3 + 3?",
+                }
+            )
+            + "\n"
+        )
+        tmp_file_path = Path(tmp_file.name)
+
+    with pytest.raises(
+        ValueError,
+        match="Sample at index 0 has invalid 'final_answer'",
+    ):
+        StaticDataset(data=tmp_file_path, min_samples=1, strict=True)
+
+    # **Test 8: JSONL-specific edge cases:
+    # Missing required field in non-strict mode**
+    with tempfile.NamedTemporaryFile(
+        mode='w', suffix='.jsonl', delete=False
+    ) as tmp_file:
+        tmp_file.write(
+            json.dumps(
+                {
+                    "question": "What is 3 + 3?",
+                }
+            )
+            + "\n"
+        )
         tmp_file_path = Path(tmp_file.name)
 
     dataset_non_strict = StaticDataset(
@@ -671,12 +871,11 @@ def test_static_dataset_init_from_list():
     invalid_data_missing = [
         {
             "question": "What is 3 + 3?",
-            "final_answer": "6",
-        }  # Missing rationale
+        }  # Missing final_answer
     ]
     with pytest.raises(
         ValueError,
-        match="Sample at index 0 has invalid 'rationale': "
+        match="Sample at index 0 has invalid 'final_answer': "
         "expected str, got <class 'NoneType'>",
     ):
         StaticDataset(data=invalid_data_missing, min_samples=1, strict=True)
@@ -730,9 +929,8 @@ def test_static_dataset_init_from_list():
             "final_answer": "Valid",
         },
         {
-            "question": "Invalid question",
-            "final_answer": "Invalid",
-        },  # Missing rationale
+            "question": "Question",
+        },  # Missing final_answer
     ]
     dataset_mixed = StaticDataset(data=mixed_data, min_samples=1, strict=False)
     assert (
@@ -818,6 +1016,197 @@ def test_static_dataset_methods():
         empty_dataset.sample()
 
 
+def test_static_dataset_getitem_slice():
+    mock_data = [
+        {
+            "question": "What is 1+1?",
+            "rationale": "Basic addition.",
+            "final_answer": "2",
+        },
+        {
+            "question": "Is the Earth round?",
+            "rationale": "Scientific consensus.",
+            "final_answer": "yes",
+        },
+        {
+            "question": "What is the capital of Japan?",
+            "rationale": "Geography knowledge.",
+            "final_answer": "Tokyo",
+        },
+        {
+            "question": "How many sides does a triangle have?",
+            "rationale": "Definition of a triangle.",
+            "final_answer": "3",
+        },
+    ]
+    dataset = StaticDataset(data=mock_data, min_samples=1, strict=True)
+
+    sliced = dataset[1:4:2]
+    assert isinstance(sliced, list), "Slicing should return a list."
+    assert len(sliced) == 2, "Slice [1:4:2] should return 2 items."
+    assert all(
+        isinstance(item, DataPoint) for item in sliced
+    ), "All items should be DataPoint instances."
+    assert (
+        sliced[0].question == "Is the Earth round?"
+    ), "Index 1 question should match."
+    assert (
+        sliced[1].question == "How many sides does a triangle have?"
+    ), "Index 3 question should match."
+    assert (
+        sliced[0] is dataset[1]
+    ), "Slicing should return references to the same objects."
+
+    # Test slice with only start and stop
+    sliced = dataset[1:3]
+    assert len(sliced) == 2, "Slice [1:3] should return 2 items."
+    assert (
+        sliced[0].question == "Is the Earth round?"
+    ), "Index 1 question should match."
+    assert (
+        sliced[1].question == "What is the capital of Japan?"
+    ), "Index 2 question should match."
+
+    # Test slice with negative indices
+    sliced = dataset[-2:]
+    assert len(sliced) == 2, "Slice [-2:] should return 2 items."
+    assert (
+        sliced[0].question == "What is the capital of Japan?"
+    ), "Index -2 question should match."
+    assert (
+        sliced[1].question == "How many sides does a triangle have?"
+    ), "Index -1 question should match."
+
+    sliced = dataset[:-1]
+    assert len(sliced) == 3, "Slice [:-1] should return 3 items."
+    assert (
+        sliced[0].question == "What is 1+1?"
+    ), "Index 0 question should match."
+    assert (
+        sliced[2].question == "What is the capital of Japan?"
+    ), "Index 2 question should match."
+
+    # Test slice with step
+    sliced = dataset[::2]
+    assert len(sliced) == 2, "Slice [::2] should return 2 items."
+    assert (
+        sliced[0].question == "What is 1+1?"
+    ), "Index 0 question should match."
+    assert (
+        sliced[1].question == "What is the capital of Japan?"
+    ), "Index 2 question should match."
+
+    # Test slice with negative step
+    sliced = dataset[3:0:-1]
+    assert len(sliced) == 3, "Slice [3:0:-1] should return 3 items."
+    assert (
+        sliced[0].question == "How many sides does a triangle have?"
+    ), "Index 3 question should match."
+    assert (
+        sliced[2].question == "Is the Earth round?"
+    ), "Index 1 question should match."
+
+    sliced = dataset[::-1]
+    assert (
+        len(sliced) == 4
+    ), "Slice [::-1] should return all 4 items in reverse."
+    assert (
+        sliced[0].question == "How many sides does a triangle have?"
+    ), "Index 3 question should match."
+    assert (
+        sliced[3].question == "What is 1+1?"
+    ), "Index 0 question should match."
+
+    # Test slicing with out-of-bound indices
+    sliced = dataset[10:20]
+    assert (
+        sliced == []
+    ), "Slice [10:20] should return an empty list when out of bounds."
+
+    sliced = dataset[-10:10]
+    assert len(sliced) == 4, "Slice [-10:10] should return all items."
+    assert (
+        sliced[0].question == "What is 1+1?"
+    ), "Index 0 question should match."
+    assert (
+        sliced[3].question == "How many sides does a triangle have?"
+    ), "Index 3 question should match."
+
+    sliced = dataset[2:100]
+    assert (
+        len(sliced) == 2
+    ), "Slice [2:100] should return items from index 2 to end."
+    assert (
+        sliced[0].question == "What is the capital of Japan?"
+    ), "Index 2 question should match."
+    assert (
+        sliced[1].question == "How many sides does a triangle have?"
+    ), "Index 3 question should match."
+
+    # Test non-integer, non-slice inputs
+    with pytest.raises(
+        TypeError, match="Indexing type <class 'str'> not supported."
+    ):
+        dataset["invalid"]
+
+    with pytest.raises(
+        TypeError, match="Indexing type <class 'float'> not supported."
+    ):
+        dataset[1.5]
+
+    with pytest.raises(
+        TypeError, match="Indexing type <class 'NoneType'> not supported."
+    ):
+        dataset[None]
+
+    with pytest.raises(
+        TypeError, match="Indexing type <class 'list'> not supported."
+    ):
+        dataset[[1, 2]]
+
+    # Test slice with step=0
+    with pytest.raises(ValueError, match="slice step cannot be zero"):
+        dataset[::0]
+
+    # Test slice with non-integer start, stop, or step
+    invalid_slice = slice("a", 3, 1)
+    with pytest.raises(
+        TypeError,
+        match="slice indices must be integers or None or have "
+        "an __index__ method",
+    ):
+        dataset[invalid_slice]
+
+    invalid_slice = slice(0, "b", 1)
+    with pytest.raises(
+        TypeError,
+        match="slice indices must be integers or None "
+        "or have an __index__ method",
+    ):
+        dataset[invalid_slice]
+
+    invalid_slice = slice(0, 3, "a")
+    with pytest.raises(
+        TypeError,
+        match="slice indices must be integers or None "
+        "or have an __index__ method",
+    ):
+        dataset[invalid_slice]
+
+    # Test slicing an empty dataset
+    empty_dataset = StaticDataset(data=[], min_samples=0, strict=True)
+    assert len(empty_dataset) == 0, "Empty dataset should have length 0."
+    assert (
+        empty_dataset[:] == []
+    ), "Full slice of empty dataset should return empty list."
+    assert (
+        empty_dataset[0:0] == []
+    ), "Zero-length slice should return empty list."
+    assert (
+        empty_dataset[1:2] == []
+    ), "Out-of-bound slice should return empty list."
+
+
 @pytest.mark.asyncio
 async def test_few_shot_generator():
     r"""Test FewShotGenerator with mocked components."""
@@ -837,10 +1226,11 @@ async def test_few_shot_generator():
     mock_agent.step.return_value = MagicMock(
         msgs=[
             MagicMock(
-                parsed={
-                    'question': 'Generated Question',
-                    'rationale': 'Generated Rationale',
-                }
+                parsed=DataPoint(
+                    question='Generated Question',
+                    rationale='Generated Rationale',
+                    final_answer="final",
+                ),
             )
         ]
     )
@@ -849,8 +1239,10 @@ async def test_few_shot_generator():
     dataset = FewShotGenerator(
         seed_dataset=mock_static_dataset,
         verifier=mock_verifier,
-        agent=mock_agent,
+        model=StubModel("Stub"),
     )
+
+    dataset.agent = mock_agent
 
     # Test generate_new method
     await dataset.generate_new(2)
@@ -887,13 +1279,15 @@ async def test_generate_new():
     ]
 
     mock_agent = MagicMock()
+    mock_agent.reset = MagicMock()
     mock_agent.step.return_value = MagicMock(
         msgs=[
             MagicMock(
-                parsed={
-                    'question': 'Generated Question',
-                    'rationale': 'Generated Rationale',
-                }
+                parsed=DataPoint(
+                    question="Generated Question",
+                    rationale="Generated Rationale",
+                    final_answer="Generated Answer",
+                )
             )
         ]
     )
@@ -901,10 +1295,11 @@ async def test_generate_new():
         MagicMock(
             msgs=[
                 MagicMock(
-                    parsed={
-                        "question": "What is 5 + 6?",
-                        "rationale": "print(5 + 6)",
-                    }
+                    parsed=DataPoint(
+                        question="What is 5 + 6?",
+                        rationale="print(5 + 6)",
+                        final_answer="11",
+                    )
                 )
             ]
         ),
@@ -912,20 +1307,22 @@ async def test_generate_new():
         MagicMock(
             msgs=[
                 MagicMock(
-                    parsed={
-                        "question": "What is 3 + 5?",
-                        "rationale": "print(2 + )",
-                    }
+                    parsed=DataPoint(
+                        question="What is 3 + 5?",
+                        rationale="print(2 + )",  # Syntax error
+                        final_answer="None",  # No valid answer
+                    )
                 )
             ]
         ),
         MagicMock(
             msgs=[
                 MagicMock(
-                    parsed={
-                        "question": "What is 7 + 8?",
-                        "rationale": "print(7 + 8)",
-                    }
+                    parsed=DataPoint(
+                        question="What is 7 + 8?",
+                        rationale="print(7 + 8)",
+                        final_answer="15",
+                    )
                 )
             ]
         ),
@@ -935,8 +1332,10 @@ async def test_generate_new():
     dataset = FewShotGenerator(
         seed_dataset=mock_static_dataset,
         verifier=mock_verifier,
-        agent=mock_agent,
+        model=StubModel("Stub"),
     )
+
+    dataset.agent = mock_agent
 
     new_datapoints = await dataset.generate_new(2)
 
@@ -997,15 +1396,17 @@ async def test_generate_new_with_max_retries():
 
     # Set up mock agent with specific output sequence
     mock_agent = MagicMock()
+    mock_agent.reset.return_value = None  # Ensures it does nothing
     mock_agent.step.side_effect = [
         # Correct
         MagicMock(
             msgs=[
                 MagicMock(
-                    parsed={
-                        "question": "What is 3 + 4?",
-                        "rationale": "print(3 + 4)",
-                    }
+                    parsed=DataPoint(
+                        question="What is 3 + 4?",
+                        rationale="print(3 + 4)",
+                        final_answer="7",
+                    )
                 )
             ]
         ),
@@ -1013,10 +1414,11 @@ async def test_generate_new_with_max_retries():
         MagicMock(
             msgs=[
                 MagicMock(
-                    parsed={
-                        "question": "What is 5 + 6?",
-                        "rationale": "print(5 + )",
-                    }
+                    parsed=DataPoint(
+                        question="What is 5 + 6?",
+                        rationale="print(5 + )",  # Syntax error
+                        final_answer="None",  # No valid answer
+                    )
                 )
             ]
         ),
@@ -1024,10 +1426,11 @@ async def test_generate_new_with_max_retries():
         MagicMock(
             msgs=[
                 MagicMock(
-                    parsed={
-                        "question": "What is 7 + 8?",
-                        "rationale": "print(7 + )",
-                    }
+                    parsed=DataPoint(
+                        question="What is 7 + 8?",
+                        rationale="print(7 + )",  # Syntax error
+                        final_answer="None",
+                    )
                 )
             ]
         ),
@@ -1035,10 +1438,11 @@ async def test_generate_new_with_max_retries():
         MagicMock(
             msgs=[
                 MagicMock(
-                    parsed={
-                        "question": "What is 9 + 10?",
-                        "rationale": "print(9 + )",
-                    }
+                    parsed=DataPoint(
+                        question="What is 9 + 10?",
+                        rationale="print(9 + )",  # Syntax error
+                        final_answer="None",
+                    )
                 )
             ]
         ),
@@ -1046,20 +1450,23 @@ async def test_generate_new_with_max_retries():
         MagicMock(
             msgs=[
                 MagicMock(
-                    parsed={
-                        "question": "What is 11 + 12?",
-                        "rationale": "print(11 + 12)",
-                    }
+                    parsed=DataPoint(
+                        question="What is 11 + 12?",
+                        rationale="print(11 + 12)",
+                        final_answer="23",
+                    )
                 )
             ]
         ),
     ]
 
+    model = StubModel("OpenAI")
+
     dataset = FewShotGenerator(
-        seed_dataset=mock_static_dataset,
-        verifier=mock_verifier,
-        agent=mock_agent,
+        seed_dataset=mock_static_dataset, verifier=mock_verifier, model=model
     )
+
+    dataset.agent = mock_agent
 
     # Expect RuntimeError due to insufficient valid
     # datapoints within retry limit
@@ -1118,13 +1525,18 @@ async def test_few_shot_generator_save_to_jsonl(tmp_path):
             )
         ]
     )
+    mock_agent.reset = MagicMock()
+
+    model = StubModel("OpenAI")
 
     # Create FewShotGenerator with mocks
     dataset = FewShotGenerator(
         seed_dataset=mock_static_dataset,
         verifier=mock_verifier,
-        agent=mock_agent,
+        model=model,
     )
+
+    dataset.agent = mock_agent
 
     # Mock data
     dataset._data = [
@@ -1199,3 +1611,356 @@ async def test_few_shot_generator_save_to_jsonl(tmp_path):
     ]
     with pytest.raises(IOError, match="No such file or directory"):
         dataset.save_to_jsonl(invalid_path)
+
+
+@pytest.mark.asyncio
+async def test_few_shot_generator_flush(tmp_path):
+    r"""Test FewShotGenerator's flush method with mocked data.
+
+    This test verifies that:
+    1. Data can be successfully flushed to a JSONL file and
+       the internal data is cleared.
+    2. The method handles empty datasets appropriately.
+    3. IO errors are caught and raised as expected.
+
+    Args:
+        tmp_path: Pytest fixture providing a temporary directory.
+    """
+    # Mock the seed dataset
+    mock_static_dataset = MagicMock(spec=StaticDataset)
+    mock_static_dataset.__len__.return_value = 5
+    mock_static_dataset.__getitem__.side_effect = lambda i: DataPoint(
+        question=f"Question {i}",
+        rationale=f"Rationale {i}",
+        final_answer=f"Answer {i}",
+    )
+
+    # Mock the verifier
+    mock_verifier = MagicMock()
+    mock_verifier.verify = AsyncMock()
+    mock_verifier.verify.return_value = MagicMock(result="Verified Answer")
+
+    # Mock the agent
+    mock_agent = MagicMock()
+    mock_agent.step.return_value = MagicMock(
+        msgs=[
+            MagicMock(
+                parsed={
+                    'question': 'Generated Question',
+                    'rationale': 'Generated Rationale',
+                }
+            )
+        ]
+    )
+    mock_agent.reset = MagicMock()
+
+    model = StubModel("OpenAI")
+
+    # Create FewShotGenerator with mocks
+    dataset = FewShotGenerator(
+        seed_dataset=mock_static_dataset,
+        verifier=mock_verifier,
+        model=model,
+    )
+
+    dataset.agent = mock_agent
+
+    # Mock data
+    dataset._data = [
+        DataPoint(
+            question="What is 2 + 2?",
+            rationale="Adding 2 and 2 gives 4.",
+            final_answer="4",
+            metadata={"created": "2025-03-12T10:00:00"},
+        ),
+        DataPoint(
+            question="What color is the sky?",
+            rationale="The sky appears blue due to Rayleigh scattering.",
+            final_answer="Blue",
+            metadata={"created": "2025-03-12T10:01:00"},
+        ),
+        DataPoint(
+            question="How many sides does a triangle have?",
+            rationale="A triangle is defined as a shape with three sides.",
+            final_answer="3",
+            metadata={"created": "2025-03-12T10:02:00"},
+        ),
+        DataPoint(
+            question="What is the capital of France?",
+            rationale="France is a country in Europe, "
+            "and its capital is well-known.",
+            final_answer="Paris",
+            metadata={"created": "2025-03-12T10:03:00"},
+        ),
+    ]
+
+    # **Test 1: Successful flush**
+    file_path = tmp_path / "test_dataset.jsonl"
+    dataset.flush(file_path)
+
+    # Verify file exists and content is correct
+    assert file_path.exists(), "JSONL file was not created"
+    with file_path.open("r", encoding="utf-8") as f:
+        lines = f.readlines()
+    assert len(lines) == 4, "Incorrect number of lines in JSONL file"
+
+    # Parse and verify all lines
+    for i, line in enumerate(lines):
+        parsed_line = json.loads(line)
+        expected_dp = dataset._data[i] if i < len(dataset._data) else None
+        # Since flush clears the data, we can't
+        # directly compare with dataset._data
+        # Instead, we check against the original data structure we set up
+        original_data = [
+            {
+                "question": "What is 2 + 2?",
+                "rationale": "Adding 2 and 2 gives 4.",
+                "final_answer": "4",
+                "metadata": {"created": "2025-03-12T10:00:00"},
+            },
+            {
+                "question": "What color is the sky?",
+                "rationale": "The sky appears blue due to "
+                "Rayleigh scattering.",
+                "final_answer": "Blue",
+                "metadata": {"created": "2025-03-12T10:01:00"},
+            },
+            {
+                "question": "How many sides does a triangle have?",
+                "rationale": "A triangle is defined as a "
+                "shape with three sides.",
+                "final_answer": "3",
+                "metadata": {"created": "2025-03-12T10:02:00"},
+            },
+            {
+                "question": "What is the capital of France?",
+                "rationale": "France is a country in Europe, "
+                "and its capital is well-known.",
+                "final_answer": "Paris",
+                "metadata": {"created": "2025-03-12T10:03:00"},
+            },
+        ]
+        expected_dp = original_data[i]
+        assert (
+            parsed_line["question"] == expected_dp["question"]
+        ), f"Question mismatch at line {i+1}"
+        assert (
+            parsed_line["rationale"] == expected_dp["rationale"]
+        ), f"Rationale mismatch at line {i+1}"
+        assert (
+            parsed_line["final_answer"] == expected_dp["final_answer"]
+        ), f"Final answer mismatch at line {i+1}"
+        assert "metadata" in parsed_line, f"Metadata missing at line {i+1}"
+        assert (
+            parsed_line["metadata"]["created"]
+            == expected_dp["metadata"]["created"]
+        ), f"Metadata mismatch at line {i+1}"
+
+    # Verify that internal data is cleared
+    assert len(dataset._data) == 0, "Internal data was not cleared after flush"
+
+    # **Test 2: Flush with empty dataset**
+    dataset._data = []
+    with pytest.raises(ValueError, match="Dataset is empty. No data to save."):
+        dataset.flush(file_path)
+
+    # **Test 3: IO Error handling with flush**
+    invalid_path = tmp_path / "nonexistent" / "test.jsonl"
+    dataset._data = [
+        DataPoint(
+            question="Test",
+            rationale="Test rationale",
+            final_answer="Test answer",
+        )
+    ]
+    with pytest.raises(IOError, match="No such file or directory"):
+        dataset.flush(invalid_path)
+
+
+# Fixture for a valid JSONL file with 4 datapoints
+# Needed for other tests
+@pytest.fixture
+def mock_jsonl_file(tmp_path: Path) -> Path:
+    r"""Creates a temporary JSONL file with four valid datapoints."""
+    file_path = tmp_path / "mock_data.jsonl"
+    datapoints = [
+        {
+            "question": "What is 2 + 2?",
+            "rationale": "Adding 2 and 2 gives 4.",
+            "final_answer": "4",
+        },
+        {
+            "question": "What color is the sky?",
+            "rationale": "The sky appears blue due to scattering.",
+            "final_answer": "Blue",
+        },
+        {
+            "question": "How many sides does a triangle have?",
+            "rationale": "A triangle has three sides.",
+            "final_answer": "3",
+        },
+        {
+            "question": "What is the capital of France?",
+            "rationale": "France’s capital is well-known.",
+            "final_answer": "Paris",
+        },
+    ]
+    with file_path.open("w", encoding="utf-8") as f:
+        for dp in datapoints:
+            json.dump(dp, f)
+            f.write("\n")
+    return file_path
+
+
+# Fixture for an invalid JSONL file
+# Needed for other tests
+@pytest.fixture
+def mock_wrong_jsonl_file(tmp_path: Path) -> Path:
+    r"""Creates a temporary JSONL file with
+    invalid data for error handling tests."""
+
+    file_path = tmp_path / "mock_wrong_data.jsonl"
+    invalid_data = [
+        '{"question": "Invalid", "rationale": "Missing final_answer"}',
+        '{"question": "Invalid", "final_answer": "Answer"}',
+        "Not a JSON object",
+    ]
+    with file_path.open("w", encoding="utf-8") as f:
+        for line in invalid_data:
+            f.write(line + "\n")
+    return file_path
+
+
+@pytest.mark.asyncio
+async def test_few_shot_generator_async(
+    mock_jsonl_file: Path, mock_wrong_jsonl_file: Path
+):
+    r"""Tests FewShotGenerator in an asynchronous context: initialization,
+    async sampling, async iteration, and error handling."""
+
+    mock_static_dataset = MagicMock(spec=StaticDataset)
+    mock_verifier = MagicMock(spec=BaseVerifier)
+    mock_model = MagicMock(spec=BaseModelBackend)
+    mock_model.model_type = "mock_model_type"
+
+    # --- Test 1: Initialization from a valid JSONL file ---
+    generator = FewShotGenerator(
+        seed_dataset=mock_static_dataset,
+        verifier=mock_verifier,
+        model=mock_model,
+        data_path=mock_jsonl_file,
+    )
+    assert (
+        len(generator._data) == 4
+    ), "Expected 4 datapoints to be loaded from JSONL"
+    assert all(
+        isinstance(dp, DataPoint) for dp in generator._data
+    ), "All loaded items should be DataPoint instances"
+    original_data = generator._data.copy()
+
+    # --- Test 2: Async sampling one datapoint ---
+    sampled_dp = await generator.async_sample()
+    assert isinstance(
+        sampled_dp, DataPoint
+    ), "Sampled item should be a DataPoint"
+    assert (
+        sampled_dp in original_data
+    ), "Sampled item should be one of the loaded datapoints"
+    assert (
+        len(generator._data) == 3
+    ), "Async sampling should reduce the number of datapoints by 1"
+
+    # --- Test 3: Asynchronous iteration (__aiter__) ---
+    collected_data = []
+    async for dp in generator:
+        collected_data.append(dp)
+        if len(collected_data) == 3:
+            break
+    assert (
+        len(collected_data) == 3
+    ), "Asynchronous iteration should yield the remaining 3 datapoints"
+    assert all(
+        isinstance(dp, DataPoint) for dp in collected_data
+    ), "All iterated items should be DataPoint instances"
+    assert sorted([dp.question for dp in collected_data]) == sorted(
+        [dp.question for dp in original_data[1:]]
+    ), (
+        "Remaining datapoints should match the "
+        "original data (minus the sampled one)"
+    )
+
+    # --- Test 4: Error handling with an invalid JSONL file ---
+    with pytest.raises(ValueError):
+        FewShotGenerator(
+            seed_dataset=mock_static_dataset,
+            verifier=mock_verifier,
+            model=mock_model,
+            data_path=mock_wrong_jsonl_file,
+        )
+
+
+def test_few_shot_generator_sync(
+    mock_jsonl_file: Path, mock_wrong_jsonl_file: Path
+):
+    r"""Tests FewShotGenerator in a synchronous context: initialization,
+    sampling, iteration, and error handling."""
+
+    mock_static_dataset = MagicMock(spec=StaticDataset)
+    mock_verifier = MagicMock(spec=BaseVerifier)
+    mock_model = MagicMock(spec=BaseModelBackend)
+    mock_model.model_type = "mock_model_type"
+
+    # --- Test 1: Initialization from a valid JSONL file ---
+    generator = FewShotGenerator(
+        seed_dataset=mock_static_dataset,
+        verifier=mock_verifier,
+        model=mock_model,
+        data_path=mock_jsonl_file,
+    )
+    assert (
+        len(generator._data) == 4
+    ), "Expected 4 datapoints to be loaded from JSONL"
+    assert all(
+        isinstance(dp, DataPoint) for dp in generator._data
+    ), "All loaded items should be DataPoint instances"
+    original_data = generator._data.copy()
+
+    # --- Test 2: Sampling one datapoint ---
+    sampled_dp = generator.sample()
+    assert isinstance(
+        sampled_dp, DataPoint
+    ), "Sampled item should be a DataPoint"
+    assert (
+        sampled_dp in original_data
+    ), "Sampled item should be one of the loaded datapoints"
+    assert (
+        len(generator._data) == 3
+    ), "Sampling should reduce the number of datapoints by 1"
+
+    # --- Test 3: Synchronous iteration (__iter__) ---
+    collected_data = []
+    for dp in generator:
+        collected_data.append(dp)
+        if len(collected_data) == 3:
+            break
+    assert (
+        len(collected_data) == 3
+    ), "Synchronous iteration should yield the remaining 3 datapoints"
+    assert all(
+        isinstance(dp, DataPoint) for dp in collected_data
+    ), "All iterated items should be DataPoint instances"
+    assert sorted([dp.question for dp in collected_data]) == sorted(
+        [dp.question for dp in original_data[1:]]
+    ), (
+        "Remaining datapoints should match the "
+        "original data (minus the sampled one)"
+    )
+
+    # --- Test 4: Error handling with an invalid JSONL file ---
+    with pytest.raises(ValueError):
+        FewShotGenerator(
+            seed_dataset=mock_static_dataset,
+            verifier=mock_verifier,
+            model=mock_model,
+            data_path=mock_wrong_jsonl_file,
+        )
