@@ -77,7 +77,8 @@ async def test_multi_step_env_lifecycle():
     assert observation.question == "Step 0"
 
     # Test first step
-    action1 = Action(llm_response="Response 1")
+    action1 = Action(index=0, llm_response="Response 1")
+    # FIXME: index set to 0 temporarily until we have batched MultiStep
     result1 = await env.step(action1)
     assert isinstance(result1, StepResult)
     assert result1.reward == 1.0
@@ -88,7 +89,8 @@ async def test_multi_step_env_lifecycle():
     mock_extractor.extract.assert_called_once_with("Response 1")
 
     # Test second step
-    action2 = Action(llm_response="Response 2")
+    action2 = Action(index=0, llm_response="Response 2")
+    # FIXME: index set to 0 temporarily until we have batched MultiStep
     result2 = await env.step(action2)
     assert isinstance(result2, StepResult)
     assert result2.reward == 1.0
@@ -112,7 +114,8 @@ async def test_multi_step_env_error_handling():
     mock_extractor.extract = MagicMock(return_value="extracted_answer")
 
     env = MockMultiStepEnv(extractor=mock_extractor, max_steps=3)
-    action = Action(llm_response="Test response")
+    action = Action(index=0, llm_response="Test response")
+    # FIXME: index set to 0 temporarily until we have batched MultiStep
 
     # **Test Case 1: Environment not set up**
     # Attempt to step without setup, expect RuntimeError
