@@ -17,7 +17,6 @@ import pytest
 
 from camel.verifiers.models import (
     VerificationOutcome,
-    VerifierInput,
 )
 from camel.verifiers.python_verifier import PythonVerifier
 
@@ -29,13 +28,15 @@ def python_verifier():
 
 
 @pytest.mark.asyncio
-async def test_python_verifier_setup_and_cleanup(python_verifier):
+async def test_python_verifiersetup_andcleanup(
+    python_verifier: PythonVerifier,
+):
     r"""Test setup and cleanup of PythonVerifier."""
-    await python_verifier._setup()
+    await python_verifier.setup(uv=True)
     assert python_verifier.venv_path is not None
     assert os.path.exists(python_verifier.venv_path)
 
-    await python_verifier._cleanup()
+    await python_verifier.cleanup()
     assert python_verifier.venv_path is None or not os.path.exists(
         python_verifier.venv_path
     )
@@ -44,82 +45,82 @@ async def test_python_verifier_setup_and_cleanup(python_verifier):
 @pytest.mark.asyncio
 async def test_python_verifier_execution_success(python_verifier):
     r"""Test successful execution of a Python script."""
-    await python_verifier._setup()
+    await python_verifier.setup(uv=True)
 
     script = 'print("Hello, World!")'
-    result = await python_verifier._verify_implementation(
-        VerifierInput(llm_response=script)
-    )
-
+    result = await python_verifier.verify(solution=script, ground_truth=None)
     assert result.status == VerificationOutcome.SUCCESS
     assert result.result == "Hello, World!"
 
-    await python_verifier._cleanup()
+    await python_verifier.cleanup()
 
 
 @pytest.mark.asyncio
 async def test_python_verifier_execution_failure(python_verifier):
     r"""Test execution failure due to syntax error."""
-    await python_verifier._setup()
+    await python_verifier.setup(uv=True)
 
     script = 'print("Hello"'  # Syntax error
     result = await python_verifier._verify_implementation(
-        VerifierInput(llm_response=script)
+        solution=script, ground_truth=None
     )
 
     assert result.status == VerificationOutcome.ERROR
     assert "SyntaxError" in result.error_message
 
-    await python_verifier._cleanup()
+    await python_verifier.cleanup()
 
 
 @pytest.mark.asyncio
 async def test_python_verifier_execution_timeout(python_verifier):
     r"""Test execution timeout handling."""
     python_verifier.timeout = 1  # Short timeout for testing
-    await python_verifier._setup()
+    await python_verifier.setup(uv=True)
 
     script = 'import time; time.sleep(5)'
     result = await python_verifier._verify_implementation(
-        VerifierInput(llm_response=script)
+        solution=script, ground_truth=None
     )
 
     assert result.status == VerificationOutcome.TIMEOUT
     assert "Execution timed out" in result.error_message
 
-    await python_verifier._cleanup()
+    await python_verifier.cleanup()
 
 
 @pytest.mark.asyncio
 async def test_python_verifier_output_mismatch(python_verifier):
     r"""Test verification when output does not match ground truth."""
-    await python_verifier._setup()
+    await python_verifier.setup(uv=True)
 
     script = 'print("Wrong output")'
     result = await python_verifier._verify_implementation(
-        VerifierInput(llm_response=script, ground_truth="Expected output")
+        solution=script, ground_truth="'Expected output'"
     )
 
     assert result.status == VerificationOutcome.FAILURE
-    assert result.error_message == "Output doesn't match ground truth"
+    assert (
+        result.error_message
+        == "Output mismatch: Wrong output != Expected output"
+    )
 
-    await python_verifier._cleanup()
+    await python_verifier.cleanup()
 
 
 @pytest.mark.asyncio
 async def test_python_verifier_correct_output_matching(python_verifier):
     r"""Test verification when output matches ground truth."""
-    await python_verifier._setup()
+    await python_verifier.setup(uv=True)
 
     script = 'print("Expected output")'
     result = await python_verifier._verify_implementation(
-        VerifierInput(llm_response=script, ground_truth="Expected output")
+        solution=script, ground_truth="'Expected output'"
     )
 
     assert result.status == VerificationOutcome.SUCCESS
     assert result.result == "Expected output"
 
-    await python_verifier._cleanup()
+    await python_verifier.cleanup()
 
 
 @pytest.mark.asyncio
@@ -128,7 +129,7 @@ async def test_python_verifier_no_venv_error():
     verifier = PythonVerifier()
     script = 'print("Hello")'
     result = await verifier._verify_implementation(
-        VerifierInput(llm_response=script)
+        solution=script, ground_truth=None
     )
 
     assert result.status == VerificationOutcome.ERROR
@@ -138,13 +139,13 @@ async def test_python_verifier_no_venv_error():
 @pytest.mark.asyncio
 async def test_python_verifier_with_numpy(python_verifier):
     r"""Test execution with numpy installed in the virtual environment."""
-    await python_verifier._setup()
+    await python_verifier.setup(uv=True)
 
     script = 'import numpy as np; print(np.array([1, 2, 3]))'
     result = await python_verifier._verify_implementation(
-        VerifierInput(llm_response=script)
+        solution=script, ground_truth=None
     )
 
     assert result.status == VerificationOutcome.SUCCESS
     assert "[1 2 3]" in result.result
-    await python_verifier._cleanup()
+    await python_verifier.cleanup()
